Refactor code by using protocol default implementations instead of the BaseMedia and BaseConvertibleMedia classes. Moved Subtitle to an internal Video class and added a new subtitles array to Video. This made it easier to add the new deleteSubtitles functionality. Moved shell command execution functions to a new Command struct with a static execute function

+++ Sources/monitr/Media/IgnoredMedia.swift
@@ -15,54 +15,43 @@ import SwiftyBeaver
 import JSON
 
 /// Management for media types that we don't care about and can just delete
-final class Ignore: BaseMedia {
-    enum SupportedExtension: String {
-        case txt; case png; case jpg; case jpeg
-        case gif; case rst; case md; case nfo
-        case sfv; case sub; case idx; case css
-        case js; case htm; case html; case url
-        case php; case md5; case doc; case docx
-        case rtf; case db
-    }
+final class Ignore: Media {
+    var plexName: String = ""
 
-    override var plexName: String {
-        return path.lastComponentWithoutExtension
-    }
-    override var finalDirectory: Path {
+    static var supportedExtensions: [String] = ["txt", "png", "jpg", "jpeg",
+                                                "gif", "rst", "md", "nfo",
+                                                "sfv", "sub", "idx", "css",
+                                                "js", "htm", "html", "url",
+                                                "php", "md5", "doc", "docx",
+                                                "rtf", "db"]
+
+    var path: Path
+    var downpour: Downpour
+
+     var finalDirectory: Path {
         return "/dev/null"
     }
 
-    required init(_ path: Path) throws {
+    init(_ path: Path) throws {
         if !path.string.lowercased().contains("sample") && !path.string.lowercased().contains(".ds_store") {
             guard Ignore.isSupported(ext: path.extension ?? "") else {
                 throw MediaError.unsupportedFormat(path.extension ?? "")
             }
         }
-        try super.init(path)
+
+        // Set the media file's path to the absolute path
+        self.path = path.absolute
+        // Create the downpour object
+        downpour = Downpour(fullPath: path)
     }
 
-    /// JSONInitializable protocol requirement
-    required init(json: JSON) throws {
-        let p = Path(try json.get("path"))
-        if !p.string.lowercased().contains("sample") && !p.string.lowercased().contains(".ds_store") {
-            guard Ignore.isSupported(ext: p.extension ?? "") else {
-                throw MediaError.unsupportedFormat(p.extension ?? "")
-            }
+    func move(to plexPath: Path, log: SwiftyBeaver.Type) throws -> Media {
+        guard path.isDeletable else {
+            throw MediaError.fileNotDeletable
         }
-        try super.init(json: json)
-    }
-
-    override func move(to plexPath: Path, log: SwiftyBeaver.Type) throws -> Ignore {
         log.verbose("Deleting ignorable file: \(path.string)")
         try path.delete()
         path = ""
         return self
-    }
-
-    override class func isSupported(ext: String) -> Bool {
-        guard let _ = SupportedExtension(rawValue: ext.lowercased()) else {
-            return false
-        }
-        return true
     }
 }
